Add many more restaurants to food.json.
The criteria for selection are admittedly somewhat nebulous, but people
will just have to settle for a somewhat subjective experience (which is
what you get if you ask a local where to eat lunch anyway).

I believe that this is all of the places that need to be added, unless
we want to start straying farther from campus (which we might not want
to do, since not everyone will have a car or similar).